feature/TASK-1_global-refactoring

+++ src/main/java/com/example/banking/repository/CardRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CardRepository extends JpaRepository<CardEntity, Integer> {
+public interface CardRepository extends JpaRepository<CardEntity, Long> {
     Optional<CardEntity> findByCardNumber(String cardNumber);
 }
